move get_ansible_version_by_issue
default triager should not depend on AnsibleVersionIndexer
